+++ lab.h
@@ -4,7 +4,7 @@
 
 struct lab_ppd {
     unsigned long long solved_bits; // max 64 labs
-    unsigned long dummy[8]; // reserved space for logic extensions
+    unsigned int dummy[8]; // reserved space for logic extensions
 
     // individual lab data follows
 
@@ -17,7 +17,7 @@ struct lab_ppd {
     unsigned char herald_talkstep; // conversation is at this point
     unsigned char graveversion; // just if we change something, we'll increase the inside version number and rearrange the graves
     unsigned char graveindex[4]; // index into the described graves. slot 0=
-    unsigned long lab2_dummy[8]; // reserved for later extensions
+    unsigned int lab2_dummy[8]; // reserved for later extensions
 
     // lab3/25: underwater
     unsigned char password1[8]; // first part of the password
@@ -25,13 +25,13 @@ struct lab_ppd {
     unsigned char prisoner_talkstep; // conversation is at this point
     unsigned char guard_talkstep; // conversation is at this point (used by guard and door)
     unsigned char align_dummy1[2];
-    unsigned long lab3_dummy[8]; // reserved for later extensions
+    unsigned int lab3_dummy[8]; // reserved for later extensions
 
     // lab4/10: gnalbgaves
-    unsigned long lab4_dummy[8]; // reserved for later extensions
+    unsigned int lab4_dummy[8]; // reserved for later extensions
 
     // lab5/15: hard life (no regenerate)
-    unsigned long lab5_dummy[8]; // reserved for later extensions
+    unsigned int lab5_dummy[8]; // reserved for later extensions
 };
 
 int count_solved_labs(int cn);
