+++ palace.c
@@ -287,15 +287,15 @@ void palace_bomb(int in, int cn) {
             sound_area(it[in].x, it[in].y, 6);
 
             m = it[in].x + it[in].y * MAXMAP;
-            if ((co = map[m].ch) && (!(ch[co].flags & CF_PLAYER) || *(unsigned long *)(it[in].drdata + 1) == ch[co].ID) && strcmp(ch[co].name, "Islena")) create_show_effect(EF_BURN, co, ticker, ticker + TICKS * 60, 250, POWERSCALE * 2);
-            if ((co = map[m + 1].ch) && (!(ch[co].flags & CF_PLAYER) || *(unsigned long *)(it[in].drdata + 1) == ch[co].ID) && strcmp(ch[co].name, "Islena")) create_show_effect(EF_BURN, co, ticker, ticker + TICKS * 60, 250, POWERSCALE * 2);
-            if ((co = map[m - 1].ch) && (!(ch[co].flags & CF_PLAYER) || *(unsigned long *)(it[in].drdata + 1) == ch[co].ID) && strcmp(ch[co].name, "Islena")) create_show_effect(EF_BURN, co, ticker, ticker + TICKS * 60, 250, POWERSCALE * 2);
-            if ((co = map[m + MAXMAP].ch) && (!(ch[co].flags & CF_PLAYER) || *(unsigned long *)(it[in].drdata + 1) == ch[co].ID) && strcmp(ch[co].name, "Islena")) create_show_effect(EF_BURN, co, ticker, ticker + TICKS * 60, 250, POWERSCALE * 2);
-            if ((co = map[m - MAXMAP].ch) && (!(ch[co].flags & CF_PLAYER) || *(unsigned long *)(it[in].drdata + 1) == ch[co].ID) && strcmp(ch[co].name, "Islena")) create_show_effect(EF_BURN, co, ticker, ticker + TICKS * 60, 250, POWERSCALE * 2);
-            if ((co = map[m + 1 + MAXMAP].ch) && (!(ch[co].flags & CF_PLAYER) || *(unsigned long *)(it[in].drdata + 1) == ch[co].ID) && strcmp(ch[co].name, "Islena")) create_show_effect(EF_BURN, co, ticker, ticker + TICKS * 60, 250, POWERSCALE * 2);
-            if ((co = map[m + 1 - MAXMAP].ch) && (!(ch[co].flags & CF_PLAYER) || *(unsigned long *)(it[in].drdata + 1) == ch[co].ID) && strcmp(ch[co].name, "Islena")) create_show_effect(EF_BURN, co, ticker, ticker + TICKS * 60, 250, POWERSCALE * 2);
-            if ((co = map[m - 1 + MAXMAP].ch) && (!(ch[co].flags & CF_PLAYER) || *(unsigned long *)(it[in].drdata + 1) == ch[co].ID) && strcmp(ch[co].name, "Islena")) create_show_effect(EF_BURN, co, ticker, ticker + TICKS * 60, 250, POWERSCALE * 2);
-            if ((co = map[m - 1 - MAXMAP].ch) && (!(ch[co].flags & CF_PLAYER) || *(unsigned long *)(it[in].drdata + 1) == ch[co].ID) && strcmp(ch[co].name, "Islena")) create_show_effect(EF_BURN, co, ticker, ticker + TICKS * 60, 250, POWERSCALE * 2);
+            if ((co = map[m].ch) && (!(ch[co].flags & CF_PLAYER) || *(unsigned int *)(it[in].drdata + 1) == ch[co].ID) && strcmp(ch[co].name, "Islena")) create_show_effect(EF_BURN, co, ticker, ticker + TICKS * 60, 250, POWERSCALE * 2);
+            if ((co = map[m + 1].ch) && (!(ch[co].flags & CF_PLAYER) || *(unsigned int *)(it[in].drdata + 1) == ch[co].ID) && strcmp(ch[co].name, "Islena")) create_show_effect(EF_BURN, co, ticker, ticker + TICKS * 60, 250, POWERSCALE * 2);
+            if ((co = map[m - 1].ch) && (!(ch[co].flags & CF_PLAYER) || *(unsigned int *)(it[in].drdata + 1) == ch[co].ID) && strcmp(ch[co].name, "Islena")) create_show_effect(EF_BURN, co, ticker, ticker + TICKS * 60, 250, POWERSCALE * 2);
+            if ((co = map[m + MAXMAP].ch) && (!(ch[co].flags & CF_PLAYER) || *(unsigned int *)(it[in].drdata + 1) == ch[co].ID) && strcmp(ch[co].name, "Islena")) create_show_effect(EF_BURN, co, ticker, ticker + TICKS * 60, 250, POWERSCALE * 2);
+            if ((co = map[m - MAXMAP].ch) && (!(ch[co].flags & CF_PLAYER) || *(unsigned int *)(it[in].drdata + 1) == ch[co].ID) && strcmp(ch[co].name, "Islena")) create_show_effect(EF_BURN, co, ticker, ticker + TICKS * 60, 250, POWERSCALE * 2);
+            if ((co = map[m + 1 + MAXMAP].ch) && (!(ch[co].flags & CF_PLAYER) || *(unsigned int *)(it[in].drdata + 1) == ch[co].ID) && strcmp(ch[co].name, "Islena")) create_show_effect(EF_BURN, co, ticker, ticker + TICKS * 60, 250, POWERSCALE * 2);
+            if ((co = map[m + 1 - MAXMAP].ch) && (!(ch[co].flags & CF_PLAYER) || *(unsigned int *)(it[in].drdata + 1) == ch[co].ID) && strcmp(ch[co].name, "Islena")) create_show_effect(EF_BURN, co, ticker, ticker + TICKS * 60, 250, POWERSCALE * 2);
+            if ((co = map[m - 1 + MAXMAP].ch) && (!(ch[co].flags & CF_PLAYER) || *(unsigned int *)(it[in].drdata + 1) == ch[co].ID) && strcmp(ch[co].name, "Islena")) create_show_effect(EF_BURN, co, ticker, ticker + TICKS * 60, 250, POWERSCALE * 2);
+            if ((co = map[m - 1 - MAXMAP].ch) && (!(ch[co].flags & CF_PLAYER) || *(unsigned int *)(it[in].drdata + 1) == ch[co].ID) && strcmp(ch[co].name, "Islena")) create_show_effect(EF_BURN, co, ticker, ticker + TICKS * 60, 250, POWERSCALE * 2);
             remove_item(in);
             destroy_item(in);
         } else if (it[in].drdata[0] == 1 && it[in].carried) {
@@ -306,7 +306,7 @@ void palace_bomb(int in, int cn) {
             it[in].drdata[0] = 1;
             it[in].sprite++;
             ch[cn].flags |= CF_ITEMS;
-            *(unsigned long *)(it[in].drdata + 1) = ch[cn].ID;
+            *(unsigned int *)(it[in].drdata + 1) = ch[cn].ID;
         }
         return;
     }
